+++ order_service/database/connection.py
@@ -3,21 +3,22 @@ from typing import AsyncIterator
 from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession, async_sessionmaker
 from sqlalchemy.pool import AsyncAdaptedQueuePool
 
-class DataBase():
-    driver = 'postgresql+asyncpg'
-    username = os.getenv('DB_USERNAME')
-    password = os.getenv('DB_PASSWORD')
-    host = os.getenv('DB_HOST')
-    port = os.getenv('DB_PORT')
-    name = os.getenv('DB_NAME')
+
+class DataBase:
+    driver = "postgresql+asyncpg"
+    username = os.getenv("POSTGRES_USER")
+    password = os.getenv("POSTGRES_PASSWORD")
+    host = os.getenv("POSTGRES_HOST")
+    port = os.getenv("POSTGRES_PORT")
+    name = os.getenv("POSTGRES_DB")
 
     def __init__(self):
         self._sessionmaker = None
         self.async_engine = None
 
     def get_db_uri(self):
-        return f'{self.driver}://{self.username}:{self.password}@{self.host}:{self.port}/{self.name}'
-    
+        return f"{self.driver}://{self.username}:{self.password}@{self.host}:{self.port}/{self.name}"
+
     def create_database(self):
         self.async_engine = create_async_engine(
             self.get_db_uri(),
@@ -26,18 +27,16 @@ class DataBase():
             pool_pre_ping=True,
             pool_size=20,
             max_overflow=10,
-            pool_timeout=30
+            pool_timeout=30,
         )
         self._sessionmaker = async_sessionmaker(
-            autocommit=False, 
-            autoflush=False,
-            bind=self.async_engine
+            autocommit=False, autoflush=False, bind=self.async_engine
         )
 
     async def session(self) -> AsyncIterator[AsyncSession]:
         if self._sessionmaker is None:
             raise Exception("База данных не инициализирована")
-        
+
         async with self._sessionmaker() as session:
             try:
                 yield session
@@ -45,10 +44,10 @@ class DataBase():
                 await session.rollback()
                 raise
 
+
 db = DataBase()
+
 
 async def get_db_session():
     async for session in db.session():
         yield session
-
-
