+++ order_service/models.py
@@ -1,5 +1,6 @@
 import enum
 
+
 class OrderStatus(enum.Enum):
     PENDING = "pending"
     PROCESSING = "processing"
