+++ src/plotva.egg-info/PKG-INFO
@@ -1,0 +1,5 @@
+Metadata-Version: 2.4
+Name: plotva
+Version: 0.1.0
+Summary: Plotva microservices platform
+Requires-Python: >=3.11
